requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dataclasses==0.6
elasticsearch==7.14.0
fake-useragent==0.1.11
firebase-admin==5.0.1
Flask==2.0.1
Flask-Sockets==0.2.1
geographiclib==1.52
geopy==2.2.0
gevent==21.1.2
gevent-websocket==0.10.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
googletransx==2.4.2
greenlet==1.1.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-socks==1.2.4
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
six==1.16.0
soupsieve==2.2.1
twint==2.1.21
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
